heredoc fully implemented
