+++ srcs/srcs_pars/ast_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ast_utils.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: enzo <enzo@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: kuru <kuru@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/16 16:09:16 by enzo              #+#    #+#             */
-/*   Updated: 2025/01/16 18:32:25 by enzo             ###   ########.fr       */
+/*   Updated: 2025/01/16 20:42:28 by kuru             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,66 +30,6 @@ bool	_parser_is_token_type_redir(t_token_type type)
 	return (type == TOK_REDIR_IN || type == TOK_REDIR_OUT
 		|| type == TOK_APPEND || type == TOK_HEREDOC);
 }
-
-static void	fill_heredoc(int fd, char *delimiter)
-{
-	char	*line;
-	char	*end_heredoc;
-
-	end_heredoc = ft_strjoin(delimiter, "\n");
-	ft_putstr_fd("heredoc> ", STDOUT_FILENO);
-	line = get_next_line(STDIN_FILENO);
-	while (line && ft_strcmp(line, end_heredoc))
-	{
-		write(fd, line, ft_strlen(line));
-		ft_putstr_fd("heredoc> ", STDOUT_FILENO);
-		free(line);
-		line = get_next_line(STDIN_FILENO);
-	}
-	free(line);
-	free(end_heredoc);
-}
-
-static char	*get_heredoc_filename(void)
-{
-	char		*filename;
-	char		*char_count;
-	static int	int_count = 0;
-
-	char_count = ft_itoa(int_count);
-	if (!char_count)
-		return (NULL);
-	int_count++;
-	filename = ft_strjoin("/tmp/heredoc_", char_count);
-	free(char_count);
-	if (!filename)
-		return (NULL);
-	if (access(filename, F_OK) == 0)
-	{
-		free(filename);
-		return (get_heredoc_filename());
-	}
-	return (filename);
-}
-
-
-static char	*heredoc_handler(char *delemiter)
-{
-	int		fd;
-	char	*heredoc_filename;
-
-	heredoc_filename = get_heredoc_filename();
-	fd = open(heredoc_filename, O_CREAT | O_WRONLY | O_TRUNC, 0644);
-	if (fd < 0)
-	{
-		free(heredoc_filename);
-		return (NULL);
-	}
-	fill_heredoc(fd, delemiter);
-	close(fd);
-	return (heredoc_filename);
-}
-
 
 t_redir	*create_redir_node(t_token *token, char *file)
 {
